requirements.txt
asgiref==3.11.0
Django==6.0
django-passwords==0.3.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2
